Submit the feedback form and render flash messages
Connects to #487

This PR wires up the client-side feedback form such that when a user submits it, one or more flash messages successfully render on the screen. It has three components:

1. Render the flash_messages partial (it was not rendering at all before so the user received no feedback, which I suspect is/was leading to duplicate form submissions);
2. Correct the selector in the client-side feedback code to point at the right div selector for flashes; and
3. Use the `.done()` callback on the AJAX request that submits the form, as `.success()` seems to no longer be defined (at least, I was seeing local errors in the JS console...)

I could not otherwise reproduce this bug, so I am hesitant to say this fixes #487 but maaaaaybe it does if the source is user confusion leading to double submits.