Refactor 'View in SearchWorks' & 'View in EarthWorks' links
Previously, we prevented the rendering of EarthWorks view links
for all content that didn't have the geo content type. We now have
a lot of scanned maps in EarthWorks that deserve to be linked to
from Purl.

The logic for deciding how and when to render these links was also
convoluted, so this PRs simplifies it somewhat and expands the
tests.