Handle missing versions in purl requests depending on format
Fixes #1098
Fixes #1099

This commit changes the behavior of the purl controller when a missing version is requested. If the requested format is HTML, redirect to the versionless page (which shows the latest/head version) with a flash message letting the user know a bogus version was requested. If the requested format is anything else (XML, JSON, etc.), return a 404 so the client/consumer knows the version is bogus.