Fetch collection membership and populate IIIF collection manifests
Ref #84

This uses the new capabilities of purl-fetcher's collections
endpoint to populate and filter a list of collection members,
and then add their manifest URLs to the items list for the
collection's manifest.

Member items are filtered based on their content type and release
status, so that only publicly released items with IIIF-displayable
content types are added. The actual rights of the object are
delegated to the viewer if the manifest is loaded.